move test to no race
